+++ compose-shadow/src/androidMain/kotlin/com/adamglin/composeshadow/SoftLayerShadowContainer.kt
@@ -1,0 +1,26 @@
+package com.adamglin.composeshadow
+
+import android.os.Build
+import android.view.View
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.ComposeView
+import androidx.compose.ui.viewinterop.AndroidView
+
+@Composable
+fun SoftLayerShadowContainer(content: @Composable () -> Unit) {
+    if (Build.VERSION.SDK_INT < Build.VERSION_CODES.P) {
+        AndroidView(
+            factory = { context ->
+                ComposeView(context).apply {
+                    setLayerType(View.LAYER_TYPE_SOFTWARE, null)
+                    setContent(content)
+                }
+            },
+            modifier = Modifier.fillMaxSize()
+        )
+    } else {
+        content()
+    }
+}
